refactor(franchise-history): extract assembly into FranchiseHistoryService
Repository reduced to raw fetches; Service owns windowByTeam merge, winpct
derivation, absent-source defaults, and FranchiseRow assembly. FranchiseRow
type relocated to FranchiseHistoryServiceInterface; View re-points import.
