+++ ibl5/classes/FranchiseHistory/FranchiseHistoryRepository.php
@@ -10,9 +10,14 @@ use League\League;
 /**
  * FranchiseHistoryRepository - Data access layer for franchise history
  *
- * Retrieves franchise history and win/loss records from the database.
+ * Retrieves raw franchise history rows from the database. All cross-source
+ * merging and derived-field computation lives in {@see FranchiseHistoryService};
+ * this class returns only the raw rows the queries produce.
  *
- * @phpstan-import-type FranchiseRow from \FranchiseHistory\Contracts\FranchiseHistoryRepositoryInterface
+ * @phpstan-import-type SummaryRow from \FranchiseHistory\Contracts\FranchiseHistoryRepositoryInterface
+ * @phpstan-import-type WindowRow from \FranchiseHistory\Contracts\FranchiseHistoryRepositoryInterface
+ * @phpstan-import-type PlayoffTotalRow from \FranchiseHistory\Contracts\FranchiseHistoryRepositoryInterface
+ * @phpstan-import-type HeatTotalRow from \FranchiseHistory\Contracts\FranchiseHistoryRepositoryInterface
  *
  * @see FranchiseHistoryRepositoryInterface For the interface contract
  * @see \BaseMysqliRepository For base class documentation
@@ -20,16 +25,14 @@ use League\League;
 class FranchiseHistoryRepository extends \BaseMysqliRepository implements FranchiseHistoryRepositoryInterface
 {
     /**
-     * @see FranchiseHistoryRepositoryInterface::getAllFranchiseHistory()
+     * @see FranchiseHistoryRepositoryInterface::getFranchiseSummaryRows()
      *
-     * @return array<int, FranchiseRow>
+     * @return list<SummaryRow>
      */
-    public function getAllFranchiseHistory(int $currentEndingYear): array
+    public function getFranchiseSummaryRows(int $currentEndingYear): array
     {
-        $fiveSeasonsAgoEndingYear = $currentEndingYear - 4;
-
-        // Query 1: All-time totals from vw_franchise_summary (no redundant ibl_team_win_loss JOIN)
-        /** @var list<array{teamid: int, team_name: string, color1: string, color2: string, totwins: int, totloss: int, winpct: string, playoffs: int, div_titles: int, conf_titles: int, ibl_titles: int, heat_titles: int}> $summaryRows */
+        // All-time totals from vw_franchise_summary (no redundant ibl_team_win_loss JOIN)
+        /** @var list<SummaryRow> $summaryRows */
         $summaryRows = $this->fetchAll(
             "SELECT ti.teamid, ti.team_name, ti.color1, ti.color2,
                     fs.totwins, fs.totloss, fs.winpct, fs.playoffs,
@@ -42,8 +45,20 @@ class FranchiseHistoryRepository extends \BaseMysqliRepository implements Franch
             League::FREE_AGENTS_TEAMID
         );
 
-        // Query 2: Rolling 5-season window from `ibl_team_win_loss` directly (avoids double materialization)
-        /** @var list<array{currentname: string, five_season_wins: int, five_season_losses: int}> $windowRows */
+        return $summaryRows;
+    }
+
+    /**
+     * @see FranchiseHistoryRepositoryInterface::getFiveSeasonWindowRows()
+     *
+     * @return list<WindowRow>
+     */
+    public function getFiveSeasonWindowRows(int $currentEndingYear): array
+    {
+        $fiveSeasonsAgoEndingYear = $currentEndingYear - 4;
+
+        // Rolling 5-season window from `ibl_team_win_loss` directly (avoids double materialization)
+        /** @var list<WindowRow> $windowRows */
         $windowRows = $this->fetchAll(
             "SELECT currentname,
                     CAST(SUM(wins) AS UNSIGNED) AS five_season_wins,
@@ -56,73 +71,22 @@ class FranchiseHistoryRepository extends \BaseMysqliRepository implements Franch
             $currentEndingYear
         );
 
-        /** @var array<string, array{five_season_wins: int, five_season_losses: int}> $windowByTeam */
-        $windowByTeam = [];
-        foreach ($windowRows as $row) {
-            $windowByTeam[$row['currentname']] = [
-                'five_season_wins' => $row['five_season_wins'],
-                'five_season_losses' => $row['five_season_losses'],
-            ];
-        }
-
-        // Bulk-fetch playoff totals and HEAT totals
-        $allPlayoffTotals = $this->getAllPlayoffTotals();
-        $allHeatTotals = $this->getAllHEATTotals();
-
-        /** @var array<int, FranchiseRow> $teams */
-        $teams = [];
-        foreach ($summaryRows as $summary) {
-            $teamName = $summary['team_name'];
-            $window = $windowByTeam[$teamName] ?? ['five_season_wins' => 0, 'five_season_losses' => 0];
-            $fiveWins = $window['five_season_wins'];
-            $fiveLosses = $window['five_season_losses'];
-            $totalGames = $fiveWins + $fiveLosses;
-            $fiveWinpct = $totalGames > 0
-                ? \BasketballStats\StatsFormatter::formatPercentage($fiveWins, $totalGames)
-                : null;
-
-            $playoffTotals = $allPlayoffTotals[$teamName] ?? ['wins' => 0, 'losses' => 0, 'winpct' => '.000'];
-            $heatTotals = $allHeatTotals[$teamName] ?? ['wins' => 0, 'losses' => 0, 'winpct' => '.000'];
-
-            $teams[] = [
-                'teamid' => $summary['teamid'],
-                'team_name' => $teamName,
-                'color1' => $summary['color1'],
-                'color2' => $summary['color2'],
-                'totwins' => $summary['totwins'],
-                'totloss' => $summary['totloss'],
-                'winpct' => $summary['winpct'],
-                'playoffs' => $summary['playoffs'],
-                'five_season_wins' => $fiveWins,
-                'five_season_losses' => $fiveLosses,
-                'totalgames' => $totalGames,
-                'five_season_winpct' => $fiveWinpct,
-                'playoff_total_wins' => $playoffTotals['wins'],
-                'playoff_total_losses' => $playoffTotals['losses'],
-                'playoff_winpct' => $playoffTotals['winpct'],
-                'heat_total_wins' => $heatTotals['wins'],
-                'heat_total_losses' => $heatTotals['losses'],
-                'heat_winpct' => $heatTotals['winpct'],
-                'heat_titles' => $summary['heat_titles'],
-                'div_titles' => $summary['div_titles'],
-                'conf_titles' => $summary['conf_titles'],
-                'ibl_titles' => $summary['ibl_titles'],
-            ];
-        }
-
-        return $teams;
+        return $windowRows;
     }
 
     /**
-     * Get aggregated playoff game wins and losses for all teams in bulk
+     * Get aggregated playoff game wins and losses for all teams in bulk (raw totals).
      *
      * Uses a single SELECT from vw_playoff_series_results with conditional aggregation
      * to compute both winner and loser game tallies without UNION ALL (avoids double materialization).
      *
-     * @return array<string, array{wins: int, losses: int, winpct: string}> Map of team name -> playoff totals
+     * @see FranchiseHistoryRepositoryInterface::getRawPlayoffTotals()
+     *
+     * @return list<PlayoffTotalRow>
      */
-    private function getAllPlayoffTotals(): array
+    public function getRawPlayoffTotals(): array
     {
+        /** @var list<PlayoffTotalRow> $rows */
         $rows = $this->fetchAll(
             "SELECT
                 team_name,
@@ -138,49 +102,25 @@ class FranchiseHistoryRepository extends \BaseMysqliRepository implements Franch
             GROUP BY team_name"
         );
 
-        /** @var array<string, array{wins: int, losses: int, winpct: string}> $result */
-        $result = [];
-        foreach ($rows as $row) {
-            /** @var array{team_name: string, total_wins: int, total_losses: int} $row */
-            $wins = $row['total_wins'];
-            $losses = $row['total_losses'];
-            $totalGames = $wins + $losses;
-            $winpct = $totalGames > 0
-                ? \BasketballStats\StatsFormatter::formatPercentage($wins, $totalGames)
-                : '.000';
-            $result[$row['team_name']] = ['wins' => $wins, 'losses' => $losses, 'winpct' => $winpct];
-        }
-
-        return $result;
+        return $rows;
     }
 
     /**
-     * Get aggregated HEAT wins and losses for all teams in bulk
+     * Get aggregated HEAT wins and losses for all teams in bulk (raw totals).
      *
-     * @return array<string, array{wins: int, losses: int, winpct: string}> Map of team name → HEAT totals
+     * @see FranchiseHistoryRepositoryInterface::getRawHeatTotals()
+     *
+     * @return list<HeatTotalRow>
      */
-    private function getAllHEATTotals(): array
+    public function getRawHeatTotals(): array
     {
+        /** @var list<HeatTotalRow> $rows */
         $rows = $this->fetchAll(
             "SELECT currentname, SUM(wins) AS total_wins, SUM(losses) AS total_losses
             FROM `ibl_heat_win_loss`
             GROUP BY currentname"
         );
 
-        /** @var array<string, array{wins: int, losses: int, winpct: string}> $result */
-        $result = [];
-        foreach ($rows as $row) {
-            /** @var array{currentname: string, total_wins: int|null, total_losses: int|null} $row */
-            $wins = $row['total_wins'] ?? 0;
-            $losses = $row['total_losses'] ?? 0;
-            $totalGames = $wins + $losses;
-            $winpct = $totalGames > 0
-                ? \BasketballStats\StatsFormatter::formatPercentage($wins, $totalGames)
-                : '.000';
-            $result[$row['currentname']] = ['wins' => $wins, 'losses' => $losses, 'winpct' => $winpct];
-        }
-
-        return $result;
+        return $rows;
     }
-
 }
